+++ backend/src/status/routes.rs
@@ -7,7 +7,7 @@ use crate::status::Status;
 async fn get_status() -> impl Responder {
     return HttpResponse::Ok().json(Status {
         version: config::VERSION.to_string(),
-        max_size: *config::LIMIT,
+        max_size: *config::LIMIT as u32,
         max_views: *config::MAX_VIEWS,
         max_expiration: *config::MAX_EXPIRATION,
         allow_advanced: *config::ALLOW_ADVANCED,
